chapter_001/src/test/java/ru/obj4j/condition/DummyBotTest.java
package ru.obj4j.condition;

import org.junit.Test;
import ru.job4j.condition.DummyBot;
import static org.hamcrest.core.Is.is;
import static org.junit.Assert.assertThat;

public class DummyBotTest {
    @Test
    public void whenGreetBot() {
        DummyBot bot = new DummyBot();
        assertThat(
                bot.answer("Привет, Бот."),
                is("Привет, умник."));
    }

    @Test
    public void whenByuBot() {
        DummyBot bot = new DummyBot();
        assertThat(
                bot.answer("Пока."),
                is("До скорой встречи."));
    }

    @Test
    public void whenUnknownBot() {
        DummyBot bot = new DummyBot();
        assertThat(
                bot.answer("Как тебя зовут?"),
                is("Это ставит меня в тупик. Спросите другой вопрос."));
    }
}
